requirements.txt
pyserial>=3.5
PySide6>=6.6.0
